requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
pymongo>=4.5.0
python-dotenv>=1.0.0
requests>=2.28.0
beautifulsoup4>=4.11.0
googlesearch-python>=1.2.3
lxml>=4.6.0
python-dateutil>=2.8.0 